+++ packages/salvi/src/salvi/web/static/assets/Monitor-BKhtGLzs.js
@@ -1,4 +1,4 @@
-import{c as k,j as e,r as g,A as U,a as w,T as ie,C as le}from"./index-CyPsxCaT.js";import{l as L,s as A,m as N,h as v,C as $,e as re,a as oe}from"./Chart-pGpwEzoq.js";/**
+import{c as k,j as e,r as g,A as U,a as w,T as ie,C as le}from"./index-DUZcnN7U.js";import{l as L,s as A,m as N,h as v,C as $,e as re,a as oe}from"./Chart-DcaN5f0E.js";/**
  * @license lucide-react v0.468.0 - ISC
  *
  * This source code is licensed under the ISC license.